Fixed - Error when binder can't convert  value for a matched property
v2.0.1 - Released

Fixes:
1 - Error when binder can't convert value for a matched property
