+++ README.md
@@ -1,4 +1,4 @@
-# ModelWrapper - Current version 2.0.0
+# ModelWrapper - Current version 2.0.1
 
 >***ModelWrapper*** is a set of tools for enriching model object manipulation, allowing all data sent to the server to be captured by WrapRequest\<TModel>, feeding the model object (TModel) and allowing all other information to be extracted and used by extended functionality. - *isilveria*.
 
@@ -26,12 +26,12 @@
 
 *Package Manager*
 ----
-> ```PM> Install-Package ModelWrapper -Version 1.1.0```
+> ```PM> Install-Package ModelWrapper -Version 2.0.1```
 
 *.NET CLI*
 ----
-> ```> dotnet add package ModelWrapper --version 1.1.0```
+> ```> dotnet add package ModelWrapper --version 2.0.1```
 
 *Packet*
 ----
-> ```> paket add ModelWrapper --version 1.1.0```
+> ```> paket add ModelWrapper --version 2.0.1```
